unify and disunify for company
