+++ packages/backend/constants/salesforce.ts
@@ -116,3 +116,65 @@ export interface SalesforceLead {
     EmailBouncedDate: Date;
     IndividualId: string;
 }
+
+export interface SalesforceCompany {
+    Id: string;
+    IsDeleted: boolean;
+    MasterRecordId: string;
+    Name?: string;
+    Type?: string;
+    ParentId?: string;
+    BillingStreet?: string;
+    BillingCity?: string;
+    BillingState?: string;
+    BillingPostalCode?: string;
+    BillingCountry?: string;
+    BillingLatitude: number;
+    BillingLongitude: number;
+    BillingGeocodeAccuracy: string;
+    BillingAddress: { [key: string]: any };
+    ShippingStreet: string;
+    ShippingCity: string;
+    ShippingState: string;
+    ShippingPostalCode: string;
+    ShippingCountry: string;
+    ShippingLatitude: number;
+    ShippingLongitude: number;
+    ShippingGeocodeAccuracy: string;
+    ShippingAddress: { [key: string]: any };
+    Phone?: string;
+    Fax: string;
+    AccountNumber: string;
+    Website?: string;
+    PhotoUrl: string;
+    Sic: string;
+    Industry?: string;
+    AnnualRevenue?: number;
+    NumberOfEmployees?: number;
+    Ownership: string;
+    TickerSymbol?: string;
+    Description?: string;
+    Rating: string;
+    Site: string;
+    OwnerId: string;
+    CreatedDate: Date;
+    CreatedById: string;
+    LastModifiedDate: Date;
+    LastModifiedById: string;
+    SystemModstamp: Date;
+    LastActivityDate?: Date;
+    LastViewedDate: Date;
+    LastReferencedDate: Date;
+    Jigsaw: string;
+    JigsawCompanyId: string;
+    CleanStatus: string;
+    AccountSource: string;
+    DunsNumber: string;
+    Tradestyle: string;
+    NaicsCode: string;
+    NaicsDesc: string;
+    YearStarted: string;
+    SicDesc: string;
+    DandbCompanyId: string;
+    OperatingHoursId: string;
+}
