+++ packages/backend/services/crm/company.ts
@@ -4,18 +4,23 @@ import { TP_ID } from '@prisma/client';
 import { CompanyService } from '../../generated/typescript/api/resources/crm/resources/company/service/CompanyService';
 import { NotFoundError } from '../../generated/typescript/api/resources/common';
 import { InternalServerError } from '../../generated/typescript/api/resources/common';
-import logError, { logInfo } from '../../helpers/logger';
+import { logInfo, logError } from '../../helpers/logger';
 import revertTenantMiddleware from '../../helpers/tenantIdMiddleware';
 import revertAuthMiddleware from '../../helpers/authMiddleware';
 import { isStandardError } from '../../helpers/error';
-import { unifyCompany, disunifyCompany, UnifiedCompany } from '../../models/unified/company';
+import { unifyObject, disunifyObject } from '../../helpers/crm/transform';
+import { UnifiedCompany } from '../../models/unified/company';
 import { PipedriveCompany, PipedrivePagination } from '../../constants/pipedrive';
+import { StandardObjects } from '../../constants/common';
+
+const objType = StandardObjects.company;
 
 const companyService = new CompanyService(
     {
         async getCompany(req, res) {
             try {
                 const connection = res.locals.connection;
+                const account = res.locals.account;
                 const companyId = req.params.id;
                 const fields = req.query.fields;
                 const thirdPartyId = connection.tp_id;
@@ -55,7 +60,16 @@ const companyService = new CompanyService(
 
                         res.send({
                             status: 'ok',
-                            result: unifyCompany({ ...company.data, ...company.data?.properties }),
+                            result: await unifyObject<any, UnifiedCompany>({
+                                obj: {
+                                    ...company.data,
+                                    ...company.data?.properties,
+                                },
+                                tpId: thirdPartyId,
+                                objType,
+                                tenantSchemaMappingId: connection.schema_mapping_id,
+                                accountFieldMappingConfig: account.accountFieldMappingConfig,
+                            }),
                         });
                         break;
                     }
@@ -67,7 +81,13 @@ const companyService = new CompanyService(
                                 authorization: `Zoho-oauthtoken ${thirdPartyToken}`,
                             },
                         });
-                        company = unifyCompany(company.data.data?.[0]);
+                        company = await unifyObject<any, UnifiedCompany>({
+                            obj: company.data.data?.[0],
+                            tpId: thirdPartyId,
+                            objType,
+                            tenantSchemaMappingId: connection.schema_mapping_id,
+                            accountFieldMappingConfig: account.accountFieldMappingConfig,
+                        });
                         res.send({ status: 'ok', result: company });
                         break;
                     }
@@ -80,7 +100,16 @@ const companyService = new CompanyService(
                                 Authorization: `Bearer ${thirdPartyToken}`,
                             },
                         });
-                        res.send({ status: 'ok', result: unifyCompany(company.data) });
+                        res.send({
+                            status: 'ok',
+                            result: await unifyObject<any, UnifiedCompany>({
+                                obj: company.data,
+                                tpId: thirdPartyId,
+                                objType,
+                                tenantSchemaMappingId: connection.schema_mapping_id,
+                                accountFieldMappingConfig: account.accountFieldMappingConfig,
+                            }),
+                        });
                         break;
                     }
                     case TP_ID.pipedrive: {
@@ -93,7 +122,16 @@ const companyService = new CompanyService(
                             }
                         );
                         const company = result.data;
-                        res.send({ status: 'ok', result: unifyCompany(company) });
+                        res.send({
+                            status: 'ok',
+                            result: await unifyObject<any, UnifiedCompany>({
+                                obj: company,
+                                tpId: thirdPartyId,
+                                objType,
+                                tenantSchemaMappingId: connection.schema_mapping_id,
+                                accountFieldMappingConfig: account.accountFieldMappingConfig,
+                            }),
+                        });
                         break;
                     }
                     default: {
@@ -112,6 +150,7 @@ const companyService = new CompanyService(
         async getCompanies(req, res) {
             try {
                 const connection = res.locals.connection;
+                const account = res.locals.account;
                 const fields = req.query.fields;
                 const pageSize = parseInt(String(req.query.pageSize));
                 const cursor = req.query.cursor;
@@ -153,7 +192,18 @@ const companyService = new CompanyService(
                         });
                         const nextCursor = companies.data?.paging?.next?.after || undefined;
                         companies = companies.data.results as any[];
-                        companies = companies?.map((c: any) => unifyCompany({ ...c, ...c?.properties }));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (c: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: { ...c, ...c?.properties },
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({
                             status: 'ok',
                             next: nextCursor,
@@ -176,7 +226,18 @@ const companyService = new CompanyService(
                         const nextCursor = companies.data?.info?.next_page_token || undefined;
                         const prevCursor = companies.data?.info?.previous_page_token || undefined;
                         companies = companies.data.data;
-                        companies = companies?.map((l: any) => unifyCompany(l));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (l: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: l,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', next: nextCursor, previous: prevCursor, results: companies });
                         break;
                     }
@@ -208,7 +269,18 @@ const companyService = new CompanyService(
                                 ? String(parseInt(String(cursor)) - companies.data?.totalSize)
                                 : undefined;
                         companies = companies.data?.records;
-                        companies = companies?.map((l: any) => unifyCompany(l));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (l: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: l,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', next: nextCursor, previous: prevCursor, results: companies });
                         break;
                     }
@@ -227,7 +299,18 @@ const companyService = new CompanyService(
                         const nextCursor = String(result.data?.additional_data?.pagination.next_start) || undefined;
                         const prevCursor = undefined;
                         const companies = result.data.data;
-                        const unifiedCompanies = companies?.map((d) => unifyCompany(d));
+                        const unifiedCompanies = await Promise.all(
+                            companies?.map(
+                                async (d) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: d,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', next: nextCursor, previous: prevCursor, results: unifiedCompanies });
                         break;
                     }
@@ -248,11 +331,18 @@ const companyService = new CompanyService(
             try {
                 const companyData = req.body as UnifiedCompany;
                 const connection = res.locals.connection;
+                const account = res.locals.account;
                 const thirdPartyId = connection.tp_id;
                 const thirdPartyToken = connection.tp_access_token;
                 const tenantId = connection.t_id;
-                const company = disunifyCompany(companyData, thirdPartyId);
-                logInfo('Revert::CREATE COMPANY', connection.app?.env?.accountId, tenantId, company);
+                const company = await disunifyObject<UnifiedCompany>({
+                    obj: companyData,
+                    tpId: thirdPartyId,
+                    objType,
+                    tenantSchemaMappingId: connection.schema_mapping_id,
+                    accountFieldMappingConfig: account.accountFieldMappingConfig,
+                });
+                console.log('Revert::CREATE COMPANY', connection.app?.env?.accountId, tenantId, company);
 
                 switch (thirdPartyId) {
                     case TP_ID.hubspot: {
@@ -363,13 +453,20 @@ const companyService = new CompanyService(
         async updateCompany(req, res) {
             try {
                 const connection = res.locals.connection;
+                const account = res.locals.account;
                 const companyData = req.body as UnifiedCompany;
                 const companyId = req.params.id;
                 const thirdPartyId = connection.tp_id;
                 const thirdPartyToken = connection.tp_access_token;
                 const tenantId = connection.t_id;
-                const company = disunifyCompany(companyData, thirdPartyId);
-                logInfo('Revert::UPDATE COMPANY', connection.app?.env?.accountId, tenantId, company, companyId);
+                const company = await disunifyObject<UnifiedCompany>({
+                    obj: companyData,
+                    tpId: thirdPartyId,
+                    objType,
+                    tenantSchemaMappingId: connection.schema_mapping_id,
+                    accountFieldMappingConfig: account.accountFieldMappingConfig,
+                });
+                console.log('Revert::UPDATE COMPANY', connection.app?.env?.accountId, tenantId, company, companyId);
 
                 switch (thirdPartyId) {
                     case TP_ID.hubspot: {
@@ -450,6 +547,7 @@ const companyService = new CompanyService(
         async searchCompanies(req, res) {
             try {
                 const connection = res.locals.connection;
+                const account = res.locals.account;
                 const fields = req.query.fields;
                 const searchCriteria: any = req.body.searchCriteria;
                 const formattedFields = (fields || '').split('').filter(Boolean);
@@ -485,7 +583,18 @@ const companyService = new CompanyService(
                             }),
                         });
                         companies = companies.data.results as any[];
-                        companies = companies?.map((c: any) => unifyCompany({ ...c, ...c?.properties }));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (c: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: { ...c, ...c?.properties },
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({
                             status: 'ok',
                             results: companies,
@@ -501,7 +610,18 @@ const companyService = new CompanyService(
                             },
                         });
                         companies = companies.data.data;
-                        companies = companies?.map((c: any) => unifyCompany(c));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (c: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: c,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', results: companies });
                         break;
                     }
@@ -515,7 +635,18 @@ const companyService = new CompanyService(
                             },
                         });
                         companies = companies?.data?.searchRecords;
-                        companies = companies?.map((c: any) => unifyCompany(c));
+                        companies = await Promise.all(
+                            companies?.map(
+                                async (c: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: c,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', results: companies });
                         break;
                     }
@@ -534,7 +665,18 @@ const companyService = new CompanyService(
                             }
                         );
                         const companies = result.data.data.items.map((item) => item.item);
-                        const unifiedCompanies = companies?.map((d: any) => unifyCompany(d));
+                        const unifiedCompanies = await Promise.all(
+                            companies?.map(
+                                async (d: any) =>
+                                    await unifyObject<any, UnifiedCompany>({
+                                        obj: d,
+                                        tpId: thirdPartyId,
+                                        objType,
+                                        tenantSchemaMappingId: connection.schema_mapping_id,
+                                        accountFieldMappingConfig: account.accountFieldMappingConfig,
+                                    })
+                            )
+                        );
                         res.send({ status: 'ok', results: unifiedCompanies });
                         break;
                     }
